% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Audpc.R
\name{Audpc}
\alias{Audpc}
\title{Calculate the Area Under the Disease Progress Curve (AUDPC)}
\usage{
Audpc(evaluation, dates, type = "absolute", na.rm = FALSE)
}
\arguments{
\item{evaluation}{Table of data of the evaluations: Data frame}

\item{dates}{Vector of dates corresponding to each evaluation}

\item{type}{relative, absolute}

\item{na.rm}{weather to remove NA values or return \code{NA} when there are only
\code{NA} values}
}
\value{
Vector with relative or absolute \acronym{AUDPC}.
}
\description{
This is a modified version of \code{\link[agricolae:audpc]{agricolae::audpc()}}, which checks for \code{NA}
values and will return \code{NA} values if specified in the argument. See
\code{help(agricolae::audpc())} for more details. Note that this function starts
with a capital \dQuote{A} to avoid namespace clashes with
\CRANpkg{agricolae}.
}
\examples{
# see examples in help(agricolae::audpc)
dates <- c(14, 21, 28) # days
# example 1: evaluation - vector
evaluation <- c(40, 80, 90)
Audpc(evaluation, dates)
dates <- c(14, 21, NA) # days
Audpc(evaluation, dates)
# agricolae::audpc(evaluation,dates)
dates <- c(14, 21, 28) # days
evaluation <- c(40, 80, NA)
Audpc(evaluation, dates)
# agricolae::audpc(evaluation,dates)
}
