% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_estimated_weather.R
\name{calc_estimated_weather}
\alias{calc_estimated_weather}
\title{Calculate estimated weather from coefficients}
\usage{
calc_estimated_weather(
  w,
  start_date = "2023-04-01",
  end_date = "2023-11-30",
  lat,
  lon,
  n_stations = 1:4,
  na.rm = FALSE
)
}
\arguments{
\item{w}{A \code{data.table} of weather coefficients, output from \code{epiphytoolR::get_weather_coefs}}

\item{start_date}{a character string of a date value indicating the first date
for in the returning data.table Must be in \acronym{ISO8601} format (YYYY-MM-DD),
\emph{e.g.} \dQuote{2020-04-26}}

\item{end_date}{a character string of a date value indicating the last date
for in the returning data.table Must be in \acronym{ISO8601} format (YYYY-MM-DD),
\emph{e.g.} \dQuote{2020-04-26}}

\item{lat}{numeric, latitude of the query coordinates where weather should be
estimated}

\item{lon}{numeric, longitude of the query coordinates where weather should be
estimated}

\item{n_stations}{integer or vector of integers indicating the number of station/s
to return from the closest (1), or 3rd closest (3) or closest five stations
(1:5). ect}

\item{na.rm}{logical, remove all weather data from stations with NA rain_fall
frequency \strong{Not Recommended}. We advise the best way is to manually remove
weather stations with NAs or correct the weather data. This argument is
available if these two options are not available to the user. Default is \code{FALSE}}
}
\value{
A \code{data.table} output of calculated on \code{get_weather_coefs.R} with the
following columns:
\emph{station} - Weather station name;
\emph{lat} - latitude;
\emph{lon} - longitude;
\emph{yearday} - integer, day of the year, see \code{data.table::yday()};
\emph{wd_rd} - numeric, mean wind direction from raw data;
\emph{wd_sd_rd} - numeric, standard deviation of wind direction from raw data;
\emph{ws_rd} - numeric, mean wind speed from raw data;
\emph{ws_sd_rd} - numeric, standard deviation of wind speed from raw data;
\emph{rain_freq} - numeric, proportional chance of rainfall on this dat 0 - 1

can be formated with \code{format_weather()}
}
\description{
Calculate estimated weather from coefficients
}
\examples{
dat <- data.frame(
  station_name = "w_STATION",
  lat = -runif(1, 15.5, 28),
  lon = runif(1, 115, 150),
  state = "SA",
  yearday = 1:365,
  wd_rw = abs(rnorm(365, 180, 90)),
  wd_sd_rw = rnorm(365, 80, 20),
  ws_rw = runif(365, 1, 60),
  ws_sd_rw = abs(rnorm(365, 10, sd = 5)),
  rain_freq = runif(365, 0.05, 0.45)
  )

  calc_estimated_weather(w = dat,
    lat = -25,
    lon = 130,
    n_stations = 1)
}
