% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fpea_emergence.R
\name{calc_fpea_emergence}
\alias{calc_fpea_emergence}
\title{Calculate field-pea emergence}
\source{
\url{https://doi.org/10.1007/978-1-4020-6065-6_9}
}
\usage{
calc_fpea_emergence(sowing_date, hemisphere = "south")
}
\arguments{
\item{sowing_date}{sowing date (date)}

\item{hemisphere}{'North' or 'South' hemishere}
}
\value{
number of degree days predicted from time of sowing to emergence
}
\description{
Function calculates the time, in degree days, between sowing and
emergence (when 50\% of sown seeds have produces at least one true leaf).
}
\details{
Formula adapted from Schoeny. et al. (2007) in European Journal of Plant
Pathology. Formula is reported to correlate to R^2 = 0.91 for pea cultivae
Cheyenne grown in the west of France
}
\examples{
calc_fpea_emergence("2021-05-10")
}
