% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_svp.R
\name{calc_svp}
\alias{calc_svp}
\title{Calculate Saturated Vapour-Pressure (SVP)}
\usage{
calc_svp(Tm, eq = "BOM")
}
\arguments{
\item{Tm}{numeric, Temperature in degrees Celsius}

\item{eq}{character, type of equation to use. Defaults to \code{"BOM"}
(Bureau of Meteorology equation, Australia). Other options are \code{"Murray"}
(Murray 1967) and \code{"Tetens"} (Tetens 1930). \code{"Sapak"} is accepted as a
deprecated alias of \code{"Tetens"}.}
}
\value{
Saturated vapour-pressure in kPa, class \CRANpkg{units}
}
\description{
Calculates saturated vapour pressure from air temperature.
By default the equation described by the Bureau of Meteorology (BOM) and the
American Meteorological Society is used. The equations described by
Murray (1967) or Tetens (1930) can also be selected via the \code{eq} argument.
}
\examples{
calc_svp(Tm = 30)
calc_svp(Tm = 30, eq = "Murray")
calc_svp(Tm = 30, eq = "Tetens")
}
\references{
Murray, F. W., 1967: On the Computation of Saturation Vapor Pressure.
\emph{J. Appl. Meteor. Climatol.}, \emph{6}, 203–204,
\doi{10.1175/1520-0450(1967)006<0203:OTCOSV>2.0.CO;2}

Tetens, O., 1930: Uber einige meteorologische Begriffe.
\emph{Zeitschrift fur Geophysik}, \emph{6}, 297–309.
}
