% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DS.R
\name{check_DS}
\alias{check_DS}
\title{Check for daylight savings}
\usage{
check_DS(x, addDay = "Oct-03", subDay = "Mar-03", verbose = FALSE)
}
\arguments{
\item{x}{a vector of times}

\item{addDay}{Month and day in the year when a hour is subtracted to start
daylight savings. format should be "Month-day" or \code{format = "\%b-\%d"}}

\item{subDay}{Month and day in the year when a hour is added to end
daylight savings. format should be "Month-day" or \code{format = "\%b-\%d"}}

\item{verbose}{Will add messages to output informing the changes to the time
vector.}
}
\value{
either nothing or a vector of times POSIXct in UTC
}
\description{
Function checks a vector of POSIXct times for continuity and duplicated times
that could be explained by daylight savings.
}
