% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_weather_warnings.R
\name{check_weather_warnings}
\alias{check_weather_warnings}
\title{Check format_weather warnings}
\usage{
check_weather_warnings()
}
\value{
lines of messages with the function, type (either a message
or warning) and the output of the message or warning
}
\description{
Print internal warnings produced during the running of format_weather
}
\examples{
f <- function(x) {
   warning("This is a warning")
   message("This is a message")
   x + x}
epiphytoolR::.capture_warnings(f(1))
check_weather_warnings()
}
