% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIV.R
\name{DIV}
\alias{DIV}
\title{Daily Infection Values}
\source{
\url{https://doi.org/10.1007/978-1-4020-6065-6_9}
}
\usage{
DIV(RH, Tm, rainfall = 0.2)
}
\arguments{
\item{RH}{relative humitiy percentage (numeric)}

\item{Tm}{daily average temperature in degrees celcius}

\item{rainfall}{daily rainfall in millimeters}
}
\value{
infection value for day i ranging between 0 (no fungal growth) and 1
(optimal growth)
}
\description{
Calculates the Temperature index for day i and hour j. Temperature index is used
to calculate the Daily Infection Values (DIV) for the model published by Shoeny et. al
(2007).
}
\details{
Formula adapted from Schoeny. et al. (2007) in European Journal of Plant
Pathology.
}
\examples{
DIV(70,20, rainfall = 0)
}
