% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIV_Tm_index.R
\name{DIV_Tm_index}
\alias{DIV_Tm_index}
\title{Daily Infection Value - temperature index}
\source{
\url{https://doi.org/10.1007/978-1-4020-6065-6_9}
}
\usage{
DIV_Tm_index(temp, rainfall = 0.2)
}
\arguments{
\item{temp}{daily average temperature in degrees Celsius}

\item{rainfall}{rainfall in mm}
}
\value{
Temperature index value (numeric), used in calculating DIV
}
\description{
Calculates the Temperature index for day i and hour j. Temperature index is
used to calculate the Daily Infection Values (DIV) for the model published
by Schoeny et. al (2007).
}
\details{
Formula adapted from Schoeny. et al. (2007) in European Journal of Plant
Pathology.
}
\examples{
DIV_Tm_index(25)
}
