% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_primary_inoculum.R
\name{est_primary_inoculum}
\alias{est_primary_inoculum}
\title{Estimate blackspot primary inoculum}
\source{
\url{https://doi.org/10.1007/978-1-4020-6065-6_9}
}
\usage{
est_primary_inoculum(degree_days, rainfall)
}
\arguments{
\item{degree_days}{sum degree days in celcius over 7 days}

\item{rainfall}{vector of daily rainfall, or maximum rainfall in a week}
}
\value{
number of new blackspot lesions per plant over 7 days
}
\description{
Estimate the amount of blackspot lesions per plant caused by
primary inoculumn, or ascospores. Formula interpreted from Schoeny. et al.
(2007) in European Journal of Plant Pathology.
}
\examples{
est_primary_inoculum(degree_days = 250,
               rainfall = 15)
}
