% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_dk.R
\name{exp_dk}
\alias{exp_dk}
\title{Exponential dispersal kernel}
\usage{
exp_dk(r, a = 1)
}
\arguments{
\item{r}{Distance from the source to the destination}

\item{a}{Scale parameter}
}
\value{
a numeric value indicating the
}
\description{
Exponential dispersal kernel function as described by Karisto et al. (2022)
}
\examples{
exp_dk(r = 0:10)
exp_dk(r = 0:10, a = 5)
}
\references{
Petteri Karisto, Frédéric Suffert and Alexey Mikaberidze. (2022). Measuring Splash
Dispersal of a Major Wheat Pathogen in the Field. \emph{PhytoFrontiers™} \strong{2}, 30-40.
}
