% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bom.R
\name{get_bom}
\alias{get_bom}
\title{Get Bureau of Meterology data}
\source{
http://reg.bom.gov.au/catalogue/data-feeds.shtml#obs-state ;
http://www.bom.gov.au/other/copyright.shtml;
http://reg.bom.gov.au/other/disclaimer.shtml
}
\usage{
get_bom(
  ftp_url,
  product_id = "IDQ60910",
  download_location,
  filetype = NULL,
  file_prefix = format(Sys.time(), format = "\%y\%m\%d_\%H\%M"),
  verbose = TRUE
)
}
\arguments{
\item{ftp_url}{character, ftp URL obtained from the BOMs website}

\item{download_location}{character, Folder location of where to download the
compressed data}

\item{file_prefix}{character, prefix applied to file name. Default is the date
and time in hours and minutes.}

\item{access_warning}{logical, default = \code{TRUE}. Elects whether to print the warning
when the function is run}

\item{state}{character, Australian state for which weather observations files
will be downloaded. Options include \code{"QLD"},\code{"NSW"},\code{"NT"},\code{"VIC"},\code{"SA"},
\code{"TAS"},\code{"WA"}}
}
\value{
character string of the download file location of the downloaded compressed
file \code{.tgz}
}
\description{
Wrapper function that accesses the Bureau of Meterology (BOM) FTP site and
download it to a specified file path.
For more information see https://www.bom.gov.au/catalogue/anon-ftp.shtml

An FTP URL is needed for this function, find this on the Bureau of Meteorology
(\acronym{BOM}) website.
Please read the copyright and disclaimer for use of the data while
you are there.
}
\examples{
\dontrun{
get_bom_observations(ftp_url = ftp://ftp.bom.gov.au/???????,
                     download_location = tempdir())
}
}
