% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bom_observations.R
\name{get_bom_observations}
\alias{get_bom_observations}
\title{Download state weather observations}
\source{
http://reg.bom.gov.au/catalogue/data-feeds.shtml#obs-state ;
http://www.bom.gov.au/other/copyright.shtml; http://reg.bom.gov.au/other/disclaimer.shtml
}
\usage{
get_bom_observations(
  ftp_url,
  download_location,
  access_warning = TRUE,
  state = "QLD",
  file_prefix = format(Sys.time(), format = "\%y\%m\%d_\%H\%M")
)
}
\arguments{
\item{ftp_url}{character, ftp url obtained from the BOMs website}

\item{download_location}{character, Folder location of where to download the
compreseed data}

\item{access_warning}{logical, default = \code{TRUE}. Elects whether to print the warning
when the function is run}

\item{state}{character, Australian state for which weather observations files
will be downloaded. Options include \code{"QLD"},\code{"NSW"},\code{"NT"},\code{"VIC"},\code{"SA"},
\code{"TAS"},\code{"WA"}}

\item{file_prefix}{character, prefix applied to file name. Default is the date
and time in hours and minutes.}
}
\value{
character string of the download file loaction of the downloaded compressed
file \code{.tgz}
}
\description{
Function downloads a compressed file to a desegnated location. The downloaded
file contains three file formats for all automated weather station, xml, axf,
and json.
An FTP url is needed for this function, find this on the Bureau of Meterology
website. Please read the copyright and disclaimer for use of the data while
you are there.
}
\examples{
\dontrun{
get_bom_observations(ftp_url = ftp://ftp.bom.gov.au/???????,
                     download_location = tempdir())
}
}
