% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_coefs.R
\name{get_weather_coefs}
\alias{get_weather_coefs}
\title{Calculate weather coefficients}
\usage{
get_weather_coefs(raw_bom_file, rolling_window = 60, meta_data)
}
\arguments{
\item{raw_bom_file}{character, file path to raw bureau of Meteorology txt file}

\item{rolling_window}{integer, number of days to summarise over a rolling window}

\item{meta_data}{data.table Bureau of Meteorology, meta-data}
}
\value{
data.table, of coefficients to estimate weather during a rainfall event.
If no rainfall data is recorded in the raw weather NULL is returned without
warning.
wd_rw, mean wind direction from rolling window;
wd_sd_rw, standard deviation of wind direction from rolling window;
ws_rw, mean wind speed from rolling window;
ws_sd_rw, standard deviation of speed from rolling window;
rain_freq, historical probability of rainfall on this day based on a rolling
window.

A \code{data.table} of coefficients to estimate weather during a rainfall event.
If no rainfall data is recorded in the raw weather NULL is returned without
warning.
\emph{station_name} - Weather station name;
\emph{lat} - latitude;
\emph{lon} - longitude;
\emph{state} - political juristiction or state;
\emph{yearday} - integer, day of the year, see \code{data.table::yday()};
\emph{wd_rd} - numeric, mean wind direction from rolling window;
\emph{wd_sd_rd} - numeric, standard deviation of wind direction from rolling window;
\emph{ws_rd} - numeric, mean wind speed from rolling window;
\emph{ws_sd_rd} - numeric, standard deviation of speed from rolling window;
\emph{rain_freq} - numeric, proportional chance of rainfall on this day 0 - 1 based
on a rolling window.
}
\description{
Calculate weather coefficients
}
\details{
get_weather_coefs uses historical bom rainfall data to determine the probability of
rainfall on each day of the year. It also summarises the mean temperatures,
wind speed and direction at the time of the rainfall.
Formally called impute_rainywind
}
\examples{
\dontrun{
library(data.table)
meta_dat <- fread("cache/bom_stations.csv")
imp_dat <-
   get_weather_coefs(raw_bom_file = "./data/Weather_data/HD01D_Data_090182_999999999959761.txt",
                    rolling_window = 60,
                    meta_data = meta_dat)}
}
