% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_diurnal.R
\name{impute_diurnal}
\alias{impute_diurnal}
\title{Impute hourly diurnal weather fluctuations.}
\usage{
impute_diurnal(
  x = 1:24,
  max_obs = 95,
  min_obs = 45,
  max_hour = 4,
  min_hour = 15,
  l_out = 24
)
}
\arguments{
\item{x}{numeric or POSIX, vector. numeric vectors indicate which hour to
return. A POSIX formatted vector can also be used and the imputed relative
humidity for the hour returned.}

\item{max_obs}{numeric, maximum daily observation of relative humidity or temperature}

\item{min_obs}{numeric, minimum daily observation of relative humidity or temperature}

\item{max_hour}{integer, hour in the day when maximum observation was made.}

\item{min_hour}{integer, hour in the day when minimum observation was made.}

\item{l_out}{integer, length out of function. \code{24} for hourly observations (default.
\code{1440} for minute. \code{48} for half hourly \strong{Still Experimental!}}
}
\value{
numeric vector equal to the of length x. An the respective hour from
the day
}
\description{
Impute hourly diurnal weather fluctuations.
}
\details{
Impute hourly fluctuations in temperature or relative humidity from daily
maximums and minimums. This function uses a sin function to estimate the
diurnal fluctuations of temperature or humidity.
}
\examples{
impute_diurnal()
impute_diurnal(Sys.time())
impute_diurnal(max_obs = 22,
               min_obs = 18)
impute_diurnal(max_hour = 3,
               min_hour = 9)
impute_diurnal(max_obs = 99,
               min_obs = 45)
impute_diurnal(max_hour = 6,
               min_hour = 14)
}
