% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_temp.R
\name{impute_temp}
\alias{impute_temp}
\title{Impute Temperatures}
\usage{
impute_temp(w, rolling_window = 40)
}
\arguments{
\item{w}{Weather data of class \code{epiphy.weather}, an output from \code{epiphytoolR::format_weather()}}

\item{rolling_window}{integer, number of hours used to determine the rolling
window}
}
\value{
Weather data of class \code{epiphy.weather}, \emph{See} \code{epiphytoolR::format_weather()}
}
\description{
Impute Temperatures
}
\details{
Impute missing temperatures using a rolling impute fill function on a \code{epiphy.weather}
class data.table. \code{epiphy.weather} class is created from \code{epiphytoolR::format_weather()}
}
\examples{
set.seed(111)
weather[round(rnorm(50,
                    mean = nrow(weather)/2,
                    sd = nrow(weather)/10)),
        temp := NA_real_]
w2 <- impute_temp(weather)
}
