% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_tm.R
\name{impute_tm}
\alias{impute_tm}
\title{Impute hourly temperatures}
\usage{
impute_tm(max_tm = 28, min_tm = 10, max_hour = 14, min_hour = 5)
}
\arguments{
\item{max_tm}{numeric, maximum daily temperature}

\item{min_tm}{numeric, minimum daily temperature}

\item{max_hour}{integer, hour in the day when maximum temperature occurred}

\item{min_hour}{integer, hour in the day when minimum temperature occurred}
}
\value{
numeric vector of length 24. An estimation for each hour of the day
}
\description{
Impute hourly temperatures
}
\details{
Impute hourly temperatures from daily maximums and minimums
}
\examples{
impute_tm()
impute_tm(max_tm = 22,
          min_tm = 18)
impute_tm(max_hour = 10,
          min_hour = 1)
}
