% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_bom_data}
\alias{make_bom_data}
\title{Mock Bureau of Meteorology weather data}
\source{
simulated
}
\usage{
make_bom_data()
}
\value{
A data frame with 8,786 rows and 15 columns:
\describe{
\item{times}{record time from weather station}
\item{temp}{hourly mean temperature in degrees}
\item{rh}{hourly mean relative humidity}
...
}
}
\description{
Simulated raw weather data from AWS weather station
}
\examples{
raw_bom <- make_bom_data()
}
