% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bom_observations.R
\name{merge_axf_weather}
\alias{merge_axf_weather}
\title{Merge BOM axf weather data}
\usage{
merge_axf_weather(
  File_compressed,
  File_axf = "IDQ60910.99123.axf",
  File_formatted = "NTamborine.csv",
  base_dir = getwd(),
  verbose = FALSE
)
}
\arguments{
\item{File_compressed}{character, file path of compressed weather file "tgz"}

\item{File_axf}{character, filename of axf weather data observation file from
bom}

\item{File_formatted}{character, filename and path to the formated file which
has previously merged data}

\item{base_dir}{character file path giving the base directory for file_formatted}

\item{verbose}{logical print extra messages to asssist debugging}
}
\value{
data.table, of merged dataset
}
\description{
@details
This function takes new BOM axf files which hold 72 hours of weather observations
}
