% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bom_observations.R
\name{merge_weather}
\alias{merge_weather}
\title{Merge 'latest' BOM weather data}
\usage{
merge_weather(
  File_compressed,
  station_file = "IDQ60910.99123.axf",
  File_formatted = "NTamborine.csv",
  base_dir = getwd(),
  verbose = FALSE
)
}
\arguments{
\item{File_compressed}{character, file path of compressed weather file "tgz"}

\item{station_file}{character, uncompressed BOM file containing 10 minute weather
observations, default is North Tamborine. Works best with \code{json} files,
but can also read \code{axf} files.}

\item{File_formatted}{character, file name and path to the formatted file for
which to store formatted data and which previously stored data will be merged
with.}

\item{base_dir}{character, weather directory which contains folders where
weather data are saved, and where \code{File_formatted} data is saved.}

\item{verbose}{logical print extra messages to assist debugging}
}
\value{
data.table, of merged dataset
}
\description{
\code{merge_weather} takes the 'latest' Bureau of Meteorology \acronym{BOM}
weather observations files which hold 72 hours of weather observations in
10-30 minute intervals and a saved earlier weather data file, row binds the
two data.tables, then overwrites the old File_formatted.csv file with the
joined data.
}
