% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_binom.R
\name{multi_var_binom}
\alias{multi_var_binom}
\title{Splash dispersal multi-binomial}
\usage{
multi_var_binom(
  row_spacing = 0.3,
  width = 1,
  px = 1,
  seeding_r = 40,
  IO = 1,
  m
)
}
\arguments{
\item{row_spacing}{Crop row spacing in meters}

\item{width}{paddock width in meters}

\item{px}{pixel size, length and width of square pixel}

\item{seeding_r}{number of plants sown in a pixel (\code{px})}

\item{IO}{number of infective sites within the pixel}

\item{m}{(optional) a matrix representing each plant and the number of
infective sites for each plant}
}
\value{
probabilty matrix where each entry gives the probabiltiy of splash
dispersal from infective sites to the respective entry
}
\description{
Generate a two dimensional probability distribution representing
the probability of spore dispersal from an inoculum source. Generates the
probability distribution with each entry representing a single plant
}
\examples{
set.seed(7)
# obtain matrix of probabilities
mvb <- multi_var_binom()
image(t(mvb))
# sample the plant (matrix cell) where the spore may drop
sample(seq_along(mvb), size = 10, prob = mvb, replace = TRUE)
# sample the cardinal coordinates (matrix cell) where the spore may drop
sam_mvb <- sample(seq_along(mvb), size = 10, prob = mvb, replace = TRUE)
data.frame(x = sapply(sam_mvb,FUN = function(x) ceiling(x/ nrow(mvb))),
           y = sapply(sam_mvb,FUN = function(x) x \%\% nrow(mvb)))

# simulate spore dispersal over 40 spread events
for(sp in 1:40){
if(sp == 1){
  m1 <- multi_var_binom()
  m0 <- matrix(0,nrow = nrow(m1),ncol = ncol(m1))
  ind1 <- sample(seq_along(m1), size = 1, prob = m1, replace = TRUE)
  m0[ind1] <- 1
}else{
  m1 <- multi_var_binom(m = m0)
  ind1 <- sample(seq_along(m1), size = 1, prob = m1, replace = TRUE)
  m0[ind1] <- m0[ind1] + 1
}
}
image(t(m1))
}
