% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_dk.R
\name{power_dk}
\alias{power_dk}
\title{Power law dispersal kernal}
\usage{
power_dk(x, a = 184.9, b = 2.07)
}
\arguments{
\item{x}{numeric, Distance in meters.}

\item{a}{numeric, scale parameter which is proportional to the strength of the
source inoculum. Defaults to parameter used for stripe rust
(\emph{Pucinnia striiformis}) dispersal by Sackett and Mundt (2005).}

\item{b}{numeric, dimensionless parameter controlling the steepness of the
gradient. Defaults to parameter used for stripe rust (\emph{Pucinnia striiformis})
dispersal by Sackett and Mundt (2005).}
}
\value{
estimated lesions per leaf
}
\description{
Dispersal kernal using the power law, as described by Sackett and Mundt (2005)
}
\examples{
power_dk(5)
power_dk(1:20)
}
\references{
https://apsjournals.apsnet.org/doi/pdf/10.1094/PHYTO-95-0983
}
