% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Simulated weather data}
\format{
\subsection{\code{weather}}{

A data frame with 8,786 rows and 15 columns:
\describe{
\item{times}{record time from weather station}
\item{temp}{hourly mean temperature in degrees}
\item{rh}{hourly mean relative humidity}
...
}
}
}
\source{
simulated
}
\usage{
weather
}
\description{
Simulated weather data from two weather stations
}
\keyword{datasets}
