% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_angle.R
\name{wind_angle}
\alias{wind_angle}
\title{Estimates Bearing Conidia Dispersed by Wind Driven Rain}
\usage{
wind_angle(mean_wind_direction, stdev_wind_direction, PSPH = 1, min_stdev = 1)
}
\arguments{
\item{mean_wind_direction}{A numeric vector representing mean wind direction
at a particular time interval}

\item{stdev_wind_direction}{Refer to standard deviation of wind_direction at
a particular time interval}

\item{PSPH}{A numeric vector, estimated from \code{.estimate_spore_discharge()}}

\item{min_stdev}{the minimum possible standard deviation of wind direction
permitted. This should reflect the turbulent effect of wind movements to
prevent std deviations of near zero when wind direction is averaged.}
}
\value{
A numeric vector giving information on the angle component of conidia
dispersed by wind driven rain.
}
\description{
\code{wind_angle()} samples from normal distribution to estimate the bearing of
spore or conidia dispersal wind.
}
\examples{
wind_angle(10, 2) # returns a single estimate
wind_angle(10, 2, PSPH = 10) # returns 10 estimates
wind_angle(15, 2, PSPH = c(5, 5)) # returns 10 estimates

}
