% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_distance.R
\name{wind_distance}
\alias{wind_distance}
\title{Wind distance dispersal parameter}
\usage{
wind_distance(
  average_wind_speed_in_fifteen_minutes,
  wind_cauchy_multiplier = 0.015,
  PSPH = 1
)
}
\arguments{
\item{average_wind_speed_in_fifteen_minutes}{A \emph{numeric} value in m/s}

\item{wind_cauchy_multiplier}{A scaling parameter to estimate a Cauchy
distribution which resembles the possible distances a conidium travels due
to wind dispersal. Defaults to \code{0.015}.}

\item{PSPH}{A numeric vector estimated from '.estimate_spore_discharge()'}
}
\value{
Numerical vector, which returns distance conidia dispersed by wind
driven rain from the source of infection
}
\description{
\code{wind_distance()} Determines distance conidia disperse, in metres, by wind
driven rain Conidia are assumed to spread from the centre of each subunit.
The destination subunit, where conidia land, could be the same subunit or
another subunit within or outside the paddock.
}
\examples{
wind_distance(10) # returns a single estimate
wind_distance(10, PSPH = 10) # returns 10 estimates
wind_distance(15, PSPH = c(5, 5)) # returns 10 estimates
}
