% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Audpc.R
\name{Audpc}
\alias{Audpc}
\title{Calculate area under the disease progress curve}
\usage{
Audpc(evaluation, dates, type = "absolute", na.rm = FALSE)
}
\arguments{
\item{evaluation}{Table of data of the evaluations: Data frame}

\item{dates}{Vector of dates corresponding to each evaluation}

\item{type}{relative, absolute}

\item{na.rm}{weather to remove NA values or return NA when there are only NA
values}
}
\value{
Vector with relative or absolute audpc.
}
\description{
Calculate area under the disease progress curve
}
\details{
This is a wrapper for \CRANpkg{agricolae} which checks for NA values and will
return NA values if specified in the argument
see \code{agricolae::audpc()} for more details.
}
\examples{
# see examples in help(agricolae::audpc)
dates <- c(14, 21, 28) # days
# example 1: evaluation - vector
evaluation <- c(40, 80, 90)
Audpc(evaluation, dates)
dates <- c(14, 21, NA) # days
Audpc(evaluation, dates)
# agricolae::audpc(evaluation,dates)
dates <- c(14, 21, 28) # days
evaluation <- c(40, 80, NA)
Audpc(evaluation, dates)
# agricolae::audpc(evaluation,dates)
}
