% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_svp.R
\name{calc_svp}
\alias{calc_svp}
\title{Calculate saturated Vapour-pressure}
\usage{
calc_svp(Tm, eq = "BOM")
}
\arguments{
\item{Tm}{Temperature in degrees Celsius}

\item{eq}{Type of equation to use while calculation VPD. defaults to the \code{BOM}
(Bureau of Meterology equation, Australia). Other options, \code{"Murray"} and
\code{"Sapak"}}
}
\value{
Saturated vapour-pressure in kPa, class \CRANpkg{units}
}
\description{
This formula uses Tetens (1930) equation as described by Murray
(1967) to calculate the saturated vapour pressure from temperature in degrees
Celsius.
}
\examples{
calc_svp(Tm = 30)
calc_svp(Tm = 30, eq = "Murray")
}
\references{
https://doi.org/10.1175/1520-0450(1967)006<0203:OTCOSV>2.0.CO;2
}
