% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_influence.R
\name{circular_influence}
\alias{circular_influence}
\title{Calculate the influence of 2 dimensional vectors}
\usage{
circular_influence(x)
}
\arguments{
\item{x}{degrees (numeric)}
}
\value{
numeric double of same length as input
}
\description{
Calculate the influence of 2 dimensional vectors
}
\details{
This function determines the circular vector influence on a 1 to -1
scale. For example if the wind is blowing on a northerly vector (0 degrees)
There is no influence reduction. If wind blows at a 90 degree angle there is
no influence on the north south plane.
}
\examples{
# Wind speed strength from a northerly wind
ws <- 5 # kph
wd <- 0 # degrees
ws * circular_influence(wd)

# Wind speed strength from a north east wind
ws <- 5 # kph
wd <- 45 # degrees
ws * circular_influence(wd)

# Wind speed strength from a south west wind
ws <- 5 # kph
wd <-225 # degrees
ws * circular_influence(wd)
}
