% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease_aug_soil.R
\name{disease_aug_soil}
\alias{disease_aug_soil}
\title{DiseaseAUGSoil - number of girdled internodes}
\source{
\url{https://doi.org/10.1094/PDIS-01-11-0077}
}
\usage{
disease_aug_soil(DNA_pg)
}
\arguments{
\item{DNA_pg}{Quantity of DNA in picograms per gram of soil of
D. pinodes, P. medicaginis var pinodella and P. koolunga.}
}
\value{
estimated number of girdled field pea internodes at the end of winter.
}
\description{
Estimates the number of girdleed field pea internodes at the end of winter from
the sum quantitative PCR blackspot pathogens, Didymella pinodes, Phoma medicaginis
var pinodella and P. koolunga. Formula described in paper linked in the DOI
below
}
\examples{
disease_aug_soil(50)
}
\seealso{
formula one \url{https://doi.org/10.1111/ppa.12044}
}
