% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiphytoolR-package.R
\docType{package}
\name{epiphytoolR-package}
\alias{epiphytoolR}
\alias{epiphytoolR-package}
\title{epiphytoolR: Epiphytotic Modeling Tools}
\description{
epiphytoolR provides functions for undertaking epiphytotic modelling and simulation studies in R. This package provides common base level functions that can be used in modelling the spatial movement of splash and wind dispersed plant disease inoculum.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paulmelloy.github.io/epiphytoolR/}
}

}
\author{
\strong{Maintainer}: Paul Melloy \email{p.melloy@uq.edu.au} (\href{https://orcid.org/0000-0003-4253-7167}{ORCID})

Other contributors:
\itemize{
  \item Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID}) [contributor]
}

}
\keyword{internal}
