% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_time_gaps.R
\name{fill_time_gaps}
\alias{fill_time_gaps}
\title{Fill time gaps in data.frame}
\usage{
fill_time_gaps(
  dat,
  t_col,
  interval = "auto",
  max_interval = 60,
  impute = FALSE
)
}
\arguments{
\item{dat}{data.frame or data.table with column of times}

\item{t_col}{character, colname of column containing times}

\item{interval}{numeric, expected time interval in minutes between times}

\item{max_interval}{numeric, maximum acceptable interval between data
observaions in minutes.}

\item{impute}{logical, not in operation, possible future functionability.
default is FALSE}
}
\value{
data.table, with extra rows containing times where there were gaps
}
\description{
Fill time gaps in data.frame
}
